Issue 18 Add a link to the hosted UI login page
Change index.html to link to the AWS hosted UI login page.

Currently I'm getting an error in the browser when I go directly on my own to the hosted UI login page, and sign up a user. I'm hoping this has to do with the fact that the login page was not linked from my site, and that this information was somehow communicated using referrer information. This commit is to test that theory.

I also have an issue where I can't log in with the created user, because it is unconfirmed.