Issue 18 Use the Email pool URL instead of the General one
This change has the Mad Liberation API call in the static site use the Cognito URL for the Email user pool, instead of the General user pool. This allows for users to be confirmed. The General user pool does not require users to enter an email address, so they could not be confirmed and could not log in after signing up.

Also, we want to get all users' email addresses anyway so that we can contact them after the game.