Add endpoints that send cookies
This adds various endpoints and classes added while experimenting with
three main topics:
  1. How to get a Lambda to send back cookies that can be read by a
script. SignedCookiesHandler.java does this successfully.
  2. What form the response from a Lambda has to take. I'm sticking with
GatewayResponse, since the getHeaders() method on the returned object
can't have duplicate header keys. To accomplish sending duplicate
set-cookie headers, as is required for sending AWS signed cookies, I'm
varying the case of "set-cookie," as recommended on the forum topic on
this problem linked in Issue 50.
  3. How to access query params from within a Lambda. You can do this
using the queryStringParameters key in the input map. You don't need to
put anything in your template to be able to do this. The query params
are always there. You can't, however, access the fragment (it seems to
be only available within browsers), so I can't have Cognito redirect to
a Lambda (which would then grab the JWT, thus keeping the JWTs out of
the browser).

madliberation-50