Use the AWS account id in the pool domain name
gh-274

This relates to finding a name for the Cognito domain.

It has to be globally unique.

A recent run, which attempted to use the User Pool id (not known until
deployment, complicating matters), failed with the error:

The domain name contains an invalid character. Domain names can only
contain lower-case letters, numbers, and hyphens. Please enter a
different name that follows this format:
^[a-z0-9](?:[a-z0-9\-]{0,61}[a-z0-9])?$

Looks like we get about 62 characters for the domain prefix, so I might
want to add a fixed-size hashing feature to @cdk-turnkey/stackname to
ensure we stay in that limit.