Remove the User Pool Domain so I can rename it
gh-274

This is based on the workaround for the exact problem described here:

https://github.com/aws/aws-cdk/issues/10062

Basically I got the error "User pool already has a domain configured."
when I tried to rename the User Pool Domain.

So here I'm deleting the user pool domain so I can re-deploy it with a
new name. It's the workaround described as a "2-stage deploy" in the
issue comments.