Dry-run save client data to a bucket
gh-274

This tries to validate the syntax for saving the user pool client secret
to S3, where backend will be able to find it. It saves non-secret data
from the Cognito user pool client, but if the data winds up in the
bucket without being printed in Actions, I'll change it so that it saves
the actual secret.