Add the unused build-frontend script
gh-274

This was part of a currently unfavored effort to communicate the Cognito
hosted UI URL to the frontend by injecting an environment variable at
build time. I am going to pursue instead the strategy of having the Log
In link on the home page link to /prod/login, and having backend's
/prod/login return a redirect to the Cognito hosted UI URL. I'm
committing scripts/build-frontend.sh, which will not be used in the
preferred setup, so I have a record of it.