Assume a non-env-specific frontend build
gh-274

Instead of having the frontend need to know env-specific details, of
which there is currently only (as far as the frontend is concerned) the
Cognito Hosted UI URL, I am going to try to keep the frontend
env-agnostic and have it pull env-specific info from the backend by
exploiting the fact that the frontend can get to the backend using
relative URLs. This is because the frontend and backend will be running
from the same domain name thanks to CloudFront.