Collect users' email on signup
gh-281

When I tried to sign up using the hosted UI, it failed. I think this is
because Cognito couldn't confirm my users, because they didn't have
email addresses collected.

I got a failure message as shown in gh-281, but when I look in the user
pool in the AWS console, the users are there with a status of
"UNCONFIRMED."

I am expecting this change to delete and re-create the user pool,
because attributes cannot change from required to not-required after
user pool creation.