Fix the ue1 region, output the param name
gh-281

I changed what I expected to be the param name that I'm conditioning an
output on, and the output didn't change. I think the ue1 stack is
deploying not in us-east-1.

I'm also outputting the param that ue1 is attempting to condition on.
SSM params that are conditioned on in a stack should always be output by
the stack.