Handle the legitimate case when params are unset
gh-281

It turns out getParameters does not populate err if the requested params
are all unset. It populates data.InvalidParameters, and leaves
data.Parameters as an empty array.

This change handles the case where Parameters is an empty array. This is
not an error case for my app--it just means that the default behavior
for the unset parameters will be used.