Remove deprecatd output domainName
gh-281

The CloudFront domainName property is deprecated in favor of
distributionDomainName, which I was already also using.

https://github.com/aws/aws-cdk/blob/349a6e2bfaa72440deb3767fb1e28e38cc4d73ef/packages/%40aws-cdk/aws-cloudfront/lib/distribution.ts#L26

The real reason for this commit is to trigger a re-build, since I have
now deleted the stack in us-west-1 that was causing name collisions on
the CloudFront Origin Request Policies. ORPs need globally unique names,
and the name was being assigned the same value (confirmed by examining
the templates for the stacks in the CloudFormation console) in the two
regions.