Lock the SES region to the webapp region
gh-281

As far as I can tell, cross-region Cognito-SES access does not work,
even when SES is in a region supported for Cognito-SES access, and even
when the sender is production-enabled.

The more important reason for this commit is that I just set the SSM
params to tell webapp to use SES for sending emails.

I have been committing, rather than clicking "re-run jobs" in GitHub
actions, to re-trigger builds, because it seems that only the console
output from the most recent run of a build for a given commit shows up.