Set up DNS for custom names
gh-293

This attempts to set it up so an SSM-specified param domainName will, if
provided, set up the necessary resources so domainName points to the
distro.

At the time of this commit, the SSM param is set up for the doug account
so that https://293-dname.dougnaphas.com will point to the app if
everything is set up correctly.