Try to get content at a specified subdomain
gh-293

My last build spun until cancelled, I believe because it was infinitely
looking for the CNAME records to validate the cert.

In addition to some minor cleanup in lib/, I'm pushing this change
because I have delegated authority over 293-dname.dougnaphas.com from
dougnaphas.com to 293-dname.dougnaphas.com. I'm hoping that will allow
CloudFormation to automatically take care of securing the cert.

I should probably add a check in bin/ to make sure that an account that
specifies a domain name can, for example, make a TXT record in the zone.