Get the existing zone by id and name
gh-293

As discussed in these issues:

https://github.com/aws/aws-cdk/issues/3663
https://github.com/aws/aws-cdk/issues/3558

you can't get an acutal Hosted Zone just with the ID. What you wind up
with when you query by ID, as I was doing prior to this commit, is some
kind of mock object that sounds like it helps with local synth,
but not with actually deploying resources.

AWS does not seem like it will fix this unfortunately.

This change tries to fetch by both name and ID.

I was getting errors like:

Error: HostedZone.fromHostedZoneId doesn't support "zoneName"

https://github.com/douglasnaphas/madliberation/runs/2062714182?check_suite_focus=true

I believe because the hosted zone I fetched was not a real hosted zone.

The docs for fromHostedZoneId
(https://docs.aws.amazon.com/cdk/api/latest/docs/@aws-cdk_aws-route53.HostedZone.html#static-fromwbrhostedwbrzonewbridscope-id-hostedzoneid),
which I was using, do say:

> Hosted zone name becomes unavailable through this query.

A CDK maintainer adds the shocking statement:

> The first two methods create use different Classes of IHostedZone that
> throw an error if you use the wrong parameter (e.g. calling getName()
> when you used fromHostedZoneId())

https://github.com/aws/aws-cdk/issues/3663#issuecomment-524914334

It is very disappointing that the CDK maintainers decided to provide an
implementation of an interface that throws an error when you try to get
its name.