Throw on db error in content
gh-278

This is just about my last idea for getting the Action to fail if the
content metadata can't be saved in Dynamo.

If this doesn't work I'm going to add an Action step that will run in
the shell and make SYNCHRONOUS calls to the AWS API to look for the
expected items, and fail if they aren't found.