Fail on script failures: frontend, itest
gh-291

My last build

https://github.com/douglasnaphas/madliberation/runs/2103993709?check_suite_focus=true

has an integration test error. I was deliberately trying to fail the
test. I had exposed the whole itest, which is not yet implemented.
However, the step and build succeeded. I'm not sure why my itest script
has ever failed when the test failed, but this should fix it.