Allow GET, HEAD, POST, DELETE
gh-291

My last build failed because the POST to /prod/room-code received a 403.
I believe this was because I had not specified the AllowedOptions for
the CloudFront distro, and the default did not include POST. This commit
specifies the allowed HTTP methods for requests to the backend.

I definitely need GET and POST, HEAD seems useful for debugging, and I
may implement DELETE some day, especially if I want to comply with
privacy rules requiring that users be able to delete their data. I don't
see a need for OPTIONS, since this site is intended to be accessed only
from its own domain name.