Restore the prod2 deployment
Deployment to the lower-level accounts went fine. I had to re-deploy to
them because I had deleted their user pools and domains, then their
stacks, then thought better of it, and restored them. I had needed to
restore their stacks because of the problem with the User Pool not
existing as describe in my last commit message.

I expect no changes to the prod2 (passover.lol) stack from this build.