Upgrade react-scripts to 5.0.0, to upgrade jest
gh-388

Jest was doing that thing where I have like "$desc" as the first arg to
the function that comes back from test.each(), and it's printing
literally "$desc" instead of resolving it. The only way I've been able
to fix this in the past is to upgrade Jest, so that's what I'm doing
here, by upgrading React Scripts.