Remove opaque cookie GSI sort key
gh-253

I want to look up the email address simply from the opaque cookie. If I
leave the sort key, I have to handle the possibility of multiple email
addresses coming back associated with the opaque cookie partition key.
If I were to specify the email address as the sort key when looking up
the opaque cookie, I would have to sanitize the email address first. The
regexes to do so are complicated.

https://stackoverflow.com/questions/201323/how-can-i-validate-an-email-address-using-a-regular-expression

I'm not positive that removing the sort key of an existing index will
work.