Return nothing on failed queries in handleJoin
gh-395

The return value from the handleJoin function is not the return value
from the Lambda Function. Even if it were, since there could be multiple
Records in an event, I do not want to fail out from the whole stream
handler invocation on one bad query. I want to still process the other
Records in the stream handler invocation's event.