Use a page-specific roster WS API
gh-395

We are going to have one WebSocket API per page where the client wants
to listen to things. It has to be one API, not one stage, because I want
different things to happen in the connect handler for each API.
Specifically, I want to be like "here's a connection for this page's WS
API." Then I want handlers to only talk to the WS APIs that they care
about, and to only be triggered by the events in DynamoDB that they care
about.