Add a third itest player who does nothing
gh-395

I thought for a moment I needed a third player to do non-WS flows to
verify that they still work. But I don't think that can be done. With
the /ws-wait and /ws-read-roster flows, players don't have the ability
to advance using the pre-Web Sockets buttons. The page will advance for
them.

I'll probably revert this right away.