Revert "Revert "Add a third itest player who..."
gh-395

This reverts commit 06db7034d83fa6b207e4c89c9c0f7e1aaff984f2.

I'm un-reverting the commit adding p3, because I want to get the p3
thing working, because I might be able to (and need to) test the pre-Web
Sockets flows by emulating mobile browser behavior in my itest. With a
mobile browser, I can't count on the tab being active when a message
arrives over the socket.