Remove styling, so I can modernize it
gh-441

I think I'm using an outdated means of styling that is incompatible with
React 18. I'm getting some test failures, the ones skipped here. I want
to deploy the app so I can debug the test failures. I realize that I
need a way of iterating on styles locally, or something to avoid
deploying to iterate on HTML and styles, but I don't have it yet.