Deploy via npm scripts
gh-443

I've wanted to make npm scripts the interface to the app for a while.
I'm doing this as part of making deploy-frontend available when running
locally, so that we can deploy frontend only from local to S3, as a
maybe-temporary compromise between full local frontend dev and full
cloud dev.