Remove typo from CloudFront Function
gh-451

Hopefully this explains why my distro has been giving the error:

503 ERROR
The request could not be satisfied.
The CloudFront function associated with the CloudFront distribution is
invalid or could not run. We can't connect to the server for this app or
website at this time. There might be too much traffic or a configuration
error. Try again later, or contact the app or website owner.
If you provide content to customers through CloudFront, you can find
steps to troubleshoot and help prevent this error by reviewing the
CloudFront documentation.