Don't export from the CloudFront Function
gh-451

Factoring the CloudFront Function out to a file resulted in the familiar
503 error saying that the Function was invalid.

I'm not sure if the old JS rules for CloudFront Function don't allow
`module.exports = handler;`. If they do, I'm not sure how to get the
Function code into CloudFront in a way that will also expose it for my
test. I need it in a unit test so that I can work on the logic for path
rewriting.