Remove AppBucket so I can rename it appBucket
gh-451

I'm on macOS with a case-insensitive file system, so I can't rename
AppBucket.ts to appBucket.ts. I want to do so because I want a function
that returns a Bucket, not a class. I want this because I want to create
the output with the Bucket name in the function, so I can cut lines out
of lib/madliberation-webapp.ts.