Add dev-deploy.sh
gh-451

My dev workflow is going to be do deploy inter-commit to the frontend
bucket. Local running is far too annoying, presently because it won't
stop trying and failing to connect to a websocket for fast refresh/hot
reloading. dev-deploy.sh finished in 37 seconds, which is fine. This may
mess up my "don't deploy if it didn't change" logic, but I can work
around that.