Revert "Set the ACL explicitly for the distro log bucket"
This reverts commit 982879a9bab4d62f08de5475c9b9ce56f8fc0571.

This didn't work. I tried it because I wanted to keep the old prod
distro logging bucket. But I made a backup of it, so I'm ready to move
on. Also I think it got blown away by my scheduled CI run, which I need
to change to only run tests.