Use key to update elements on page change
The selected page, used as defaultValue for the select and pagination
elements I've added recently, is not updating when you go to a new page
using anything but that particular element.

That is, if I go to the URL and type in a new hash value, the right page
displays, but the page showing as selected in the select and the
pagination do not update.

Trying to fix this with the key prop, which is normally for lists.

Also adding a key to the option items, which more clearly need it.

https://stackoverflow.com/questions/30792526/defaultvalue-change-does-not-re-render-input