Disable Plan Seder if there is no logged-in user
gh-568

Because my Plan a Seder button renders as an <a> element, I don't have
access to the "disabled" form control attribute.

I want to see what the HTML looks like when the button is disabled.

This says there's not a portable way to check for a disabled link:

https://stackoverflow.com/questions/10276133/how-to-disable-html-links

ChatGPT says to use:

```
link.getAttribute('aria-disabled') === 'true' ||
link.hasAttribute('disabled')
```