Remove duplicate useQuery
That was in there by mistake. I'm wondering if that was the cause for
the strange behaviors I was seeing: the page re-rendering constantly,
to the point where I couldn't even type more than one character in the
the boxes without it getting blanked out; the fetches firing and
re-firing constantly; and the XPath for 'text contains <nickname>' never
being found, despite being plainly visible on the page.