Don't require answers_map in NewImage in filter
gh-577

I think something may be wrong with my event filter for tying the
DynamoDB stream to the submit handler.

The submit handler doesn't appear to ever be firing.

I can't find any docs on whether the commented-out line here is the
right way to say "match if the new image has a map called answers_map,"
so I'm cutting it out.

This should broaden the filter so every participant item modification
causes the submit handler to fire.

This should be OK, since (1) every participant item modification
probably (definitely?) is an answer submission, and (2) I'm debouncing
on the frontend, so frequent messages should be ok.